+++ cmd/checktools.go
@@ -6,9 +6,6 @@ package cmd
 
 import (
 	"fmt"
-	"log"
-	"bytes"
-	"os/exec"
 	"github.com/spf13/cobra"
 )
 
@@ -44,15 +41,4 @@ var checktoolsCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(checktoolsCmd)
-}
-func execShellReturnStrings(command string, args ...string) (string, string, error) {
-	var outb, errb bytes.Buffer	
-	k := exec.Command(command, args...)
-	k.Stdout = &outb
-	k.Stderr = &errb
-	err := k.Run()
-	if err != nil {
-		log.Println("Error executing command: %v", err)
-	}
-	return outb.String(), errb.String(), err
 }
